requirements.txt
asgiref==3.7.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
confection==0.1.0
cymem==2.0.7
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.9
mysqlclient==2.2.0
networkx==3.1
numpy==1.25.1
packaging==23.1
pathlib==1.0.1
pathy==0.10.2
preshed==3.0.8
pydantic==1.10.11
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
smart-open==6.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
sympy==1.12
thinc==8.1.10
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wasabi==1.1.2
